requirements.txt
discord
python-telegram-bot
